⬆️ [Plugin CustomFields Sections] Compatibilité avec Redmine 5.1.1
