+++ init.rb
@@ -1,9 +1,4 @@
-require_dependency 'redmine_custom_fields_sections/hooks'
-
-Rails.application.config.to_prepare do
-  require_dependency "redmine_custom_fields_sections/project_custom_field_patch"
-  require_dependency "redmine_custom_fields_sections/custom_fields_helper_patch"
-end
+require_relative 'lib/redmine_custom_fields_sections/hooks'
 
 Redmine::Plugin.register :redmine_custom_fields_sections do
   name "Redmine Custom Fields Sections plugin"
