+++ lib/redmine_custom_fields_sections/models/project_custom_field_patch.rb
@@ -1,0 +1,15 @@
+require_dependency "project_custom_field"
+
+module RedmineCustomFieldsSections::Models
+  module ProjectCustomFieldPatch
+    def self.included(base)
+      base.class_eval do
+        belongs_to :section, class_name: "CustomFieldsSection", optional: true
+
+        safe_attributes("section_id")
+      end
+    end
+  end
+end
+
+ProjectCustomField.send(:include, RedmineCustomFieldsSections::Models::ProjectCustomFieldPatch)
